Refactor: Improve CQRS structure, Integrate AutoMapper and ILogger
The CQRS structure was improved by refactoring the 'Update', 'Get all', 'Get by Id', and 'Create' commands & queries in the Activities section to integrate AutoMapper and ILogger. These classes were renamed following the standard convention of appending 'Command' or 'Query'. This enhances code readability and eases future maintenance. The 'Activity' controller was updated to utilize these refactored commands and queries. A DependencyInjection class was introduced in the Application layer for service registration, decluttering the startup classes.