Remove redundant ApplicationUser class and corresponding documentations. Add JwtTokenService for managing JWT tokens.
The ApplicationUser class and corresponding documentations were removed because they were determined to be redundant and not used anywhere. Plus, JwtTokenService was added for creating and handling JWT tokens. The primary function of this class is to create the JWT payload, sign the JWT using the key from appsettings.json, and then return the token. This is part of a larger effort to secure the application by implementing JWT authentication and authorization. Major changes in the code also lead to adjustments to the application's project structure seen in .csproj files. Several classes were updated accordingly due to these changes. New files for added functionalities (like TokenModel, RegisterDTO, LoginDTO) were created to support the jwt tokens. Unit Test for GetActivityById was updated as well.