Implement JWT authentication, remove unused ApplicationUser class
The IApplicationUser and the ApplicationUser classes have been removed due to redundancy as they were not used in the application. Instead, JWT authentication has been implemented using the JwtTokenService, which creates and manages JWT tokens. Responsibilities of the JwtTokenService include signing the JWT with a key from appsettings.json and returning the token. This change is a step to secure the application by establishing JWT authentication and authorization rules. Other necessary updates in various parts of the application have been made to align with this change. New classes such as TokenModel, RegisterDTO, and LoginDTO were added to support the jwt tokens. Unit test for GetActivityById was also updated accordingly.