Fixed an issue where JWT Token did not work.Refactor authentication to improve security and functionality
This is due to the order in which authentication is added, and the authentication method added later will overwrite the first addition. If Bearer is added before Identity, the default AuthenticationSchemes will be Identity.

Authentication has been refactored to use JWT Bearer instead of Cookie Authentication. A JWT Token is now stored in HttpOnly Cookie minimizing the risk of XSS attacks. Changes to the handling and management of Cookies have been implemented across the application. Refresh token functionality has been enhanced and is now stored as a separate cookie. The 'register' and 'login' methods in the AccountController now return an AccountResponseDTO for increased flexibility in the data returned on successful user registration or login. The ActivitiesController now requires authorization for improved security. Miscellaneous updates to DTOs and Services have been made to support these changes.