Extend GetActivityByIdQuery to include Attendees
Integrated IUserService into GetActivityByIdQueryHandler and extended the query to fetch related Attendees when retrieving an Activity. The ActivityAttendeeDTO was refined to include UserId for proper identification. Corresponding unit tests were updated to accommodate the new changes as well. The inclusion of Attendees enables a more comprehensive data pull in one go, enhancing performance and user