Inherit AuditableEntity in UserFollowing and Photo classes
In this commit, both UserFollowing and Photo classes have been updated to inherit from AuditableEntity. In addition to this, a GUID has been added as the ID for the Photo class, replacing the previous string ID. This change aligns with adapting the database schema to support entity auditing and GUID identifiers.