Add UI components, update error messages and organize client app files
Two User Interface (UI) components namely SearchComponent and MainPage, were added. Several notable changes were made in the application error message. The string 'Error saving to the database: {ExMessage}' was replaced with 'ErrorMessage saving to the database: {ExMessage}'. Furthermore, the organization of the files was improved in the client application. Modifications were done as well in the BaseUrl, GetActivities and GetActivity functions of the APIs.