Refactor domain events and update entity models
This commit renames the `CreateActivityDomainEvent` to `ActivityCreatedDomainEvent` for consistency. It also introduces new events related to activities, comments, attendees, photos, and follows. Furthermore, it updates entities classes encapsulating add and remove operations within attendee and comment models. In the `Activity` class, the instantiation of a new event is updated in accordance with the renaming.