Modify photo and activities management, enhance code syntax
Separate the UpLoadPhoto method and add UploadActivityPhotoCommand to upload photos for activities. Amend the RemovePhotoAsync function in PhotoService.cs to protect the main photo. Besides, modify the enum parsing method and confirmation of text field in the GetPaginatedActivitiesWithAttendeesQuery .cs. Lastly, optimize the coding practice in various files and change the page layout in PageContainer.tsx.