Update authentication logic and improve JWT token management
The authentication methods have been revised in the AccountController to include refreshing tokens and checking when a user is already logged in. The `GetPrincipalFromJwtToken` method in JwtTokenService has been updated to validate the token's lifetime. A new parameter `validateLifetime` was introduced into `GetPrincipalFromJwtToken` method to set the token's validity based on given requirements. Furthermore, account-related processes are now improved, with checks implemented for overall user conformity.