Update UI elements, fix issues and improve code cleanliness
This commit includes various changes mostly focused on the ClientApp. It introduces improvements to SignUpForm and LoginForm in terms of UI and fixed some functionality. Also, optimization has been done in the mentions of theme attributes. An ActivityFilter component has been added for filter functionality in activities and updated the way activities are displayed with the addition of ActivitiesList and ActivitiyItem components. Added some checks for token management in account related actions and moved around some elements in main.tsx file for better organization. Made changes to improve code cleanliness in different areas and updated the settings and configs.