Update photo service and add user card
Refactored the PhotoService class to handle adding a new main photo and replacing the old one. Also, the UserCard.tsx file was added for user profile representation. The Dto classes were renamed (DTO to Dto) for consistency. The comments now are ordered by their creation date and not in reverse order.