Revamp user and activity features
Various changes were made to the user and activity features. This includes updated methods for following a user, managing profile and activity photos, and general UI enhancements. Fields were also added in the userProfile interface for the followers and following attributes. Changes were made to the react components/user card as well, allowing for profile photo uploads and updates.