Refactor pagination and filtering logic in activities
Pagination parameters and filters are no longer pulled from the URL location within the `useGetPaginatedActivitiesQuery` file, but are passed directly to the function. This improves modularity and maintainability of the code, because each component is now only responsible for its particular task. Additionally, a new useFilters hook is created to generate filters from URL searchParams, which is then used in HomePage.tsx to feed into the `useGetPaginatedActivitiesQuery` function.