Enhance search, filter functionalities and add a profile update feature
In this commit, several updates have been made to improve the search, filter and profile functionalities in the application. The activity search component in the UI now navigates the user to the activity display page whenever a user triggers the built-in search function. Various filters (by categories, by hosts, by attendees) have been added to the activity query to help users narrow down their search. A feature to update users' profiles has been added along with a new profile form. Lastly, checks were added to ensure that the user is either authenticated or that unique values are used when updating the user profile.