Implement activity cancellation feature and refactor code
Added a functionality to cancel an activity in backend and frontend. The process includes adding a new command for activity cancellation and an endpoint to handle the HTTP request. The infrastructure for the project was restructured, moving database configuration to a separate folder under 'Data'. The models were also updated to reflect the cancellation status of an activity. Minor UI fixes were done. The Follower and Following pages were also updated.