Refactor notification system and enhance following function
The notification system has been restructured to improve its functioning. A "Default" NotificationType has been added and the implementation of GetPaginatedNotificationQuery has been improved. Modifications were made to the handling of new activity creation and activity cancellation events. The CQRS side was also updated to handle changes in the way notifications are created. In conjunction with these changes, the frontend was refactored for better integration with the backend.