Modify UI design and attendants handling in notifications
Modified UI design for Notifications and Activity containers, along with changes to Attendee handling in the code. The 'ListContainer' max width is expanded, altered message structure in 'ActivityCreatedDomainEventHandler', added functionalities in 'DetailComments', 'CommentInput' and 'NavBar'. Created 'AttendeesPage' and revised the routing in 'Router'. Minor modifications were made to the UI elements like colors and shapes.