Add Docker workflow and update logging settings
This commit includes a new Docker workflow file for building and pushing the Docker image. The logging settings in the appsettings.Production.json file were also updated to write logs to both console and file for better monitoring. Additionally, the environment setting was changed from development to production to reflect the current working environment.