Implement Activity reactivation feature
Added a functionality to reactivate activities that were previously cancelled. This included creating a new HTTP Put endpoint in the ActivitiesController to handle reactivation requests, creating a new 'ReactiveActivityCommand' class in the application layer, adding a 'Reactive' method in the Activity entity and integrating frontend changes. Altered some existing methods and classes names to clarify their usage and changed the ASPNETCORE_ENVIRONMENT variable to "Development" for local testing.