Refactor codebase and update Docker configurations
This commit includes major changes across the codebase. The Docker configurations have been updated to support a Multi-Stage build process and to switch from SQL Server to PostgreSQL. Also, a random Australian phone number generator has been added to the User data seed method. In the frontend, the BASE_URL for the SignalR Hub connection in several files is now a fallback string to prevent any undefined errors. Additionally, a login prompt has been added in the CommentInput component when the user is not logged in.